backend/requirements.txt
Flask
PyJWT
firebase-admin
flask-cors
